backend/requirements.txt
anthropic
Deprecated
google-genai
flask
flask_cors
nltk
numpy
openai
requests
scikit-learn
torch
transformers
